drm/i915: add GTT huge page support
The hw adds supports for 64KB, 2M and 1G pages on gen9+ for the PPGTT.
For 64KB pages nothing really changes, only how we walk the PTE's, but for
2M the pdpe's become the PTE, and for 1G pages the pml4e's become the
PTE. When binding an object to the ppgtt, we can now pass a page_sz option
from the gem object, which will setup the PTE's in the usual 4K, but now
also. Currently nothing uses this, but in the next patch we pipe support
for gem objects backed by stolen memory. Where the idea will be to use 64K
pages for gem objects >= 64, and likewise for 2M and 1G.

Signed-off-by: Matthew Auld <matthew.auld@intel.com>