drm/i915: introduce gtt page size
In preparation for supporting huge gtt pages for the ppgtt, we introduce
the gtt_page_size member for gem objects.  We fill in the gtt page size
by scanning the sg table to determine the max page size which satisfies
the alignment for each sg entry. Currently the only platform supported
page size is 4K.

Signed-off-by: Matthew Auld <matthew.auld@intel.com>