drm/i915/selftests: add fast option to igt_write_huge
For our smoke tests, we really don't need to write to some dword for
every single page in an object, nor do we need to enumerate lots and
lots of GTT addresses. Instead prefer selecting a single random GTT
offset for the object, and selecting some random 4M(max) chunk for the
dword writes. Over repeated CI runs we should still catch issues, and we
still have the exhaustive subtest which uses the slow path.

Signed-off-by: Matthew Auld <matthew.auld@intel.com>