drm/i915/dg1: Eviction logic
When an object is pinning, get_pages() is called to allocate memory
on a region, if memory pages are not availabe, the region eviction
is triggered to find other objects on the same region to be evicted,
the selected object is passed to put_pages() call to free the
memory pages, before freeing the pages, whether to swap out first
to system memory depends on if object is marked as WILLNEED.

After swapped-out, the object is treated as if it does not have any
page allocated for it.

Similarly, when an object is pinning, memory pages are allocated from
a region, then the object is checked if it had been swapped out before,
if yes, swap-in the pages contents into the allocated memory pages.

For this initial swapping code, i915_gem_object_memcpy() is used to
copy pages.

Signed-off-by: CQ Tang <cq.tang@intel.com>