HAX drm/i915: Work around the selftest timeline lock splat workaround
There is a dirty hack to work around a lockdep splat because incorrect
ordering of selftest timeline lock against other locks. However, some
selftests recently started to use the same nesting level as the workaround
and thus introduced more splats. Add a workaround to the workaround making
some selftests aware of the workaround.

Signed-off-by: Thomas Hellström <thomas.hellstrom@intel.com>
Cc: Mattew Auld <matthew.auld@intel.com>