drm/i915/ttm: Initialize the ttm device and memory managers.
Temporarily remove the buddy allocator and related selftests
and hook up the TTM range manager for i915 regions. In order to support
some of the mock region-related selftests, we need to be able to
initialize the TTM range-manager standalone without a struct ttm_device
which implies some slight additions to the TTM api.

Also modify the mock region selftests somewhat to account for a
fragmenting manager.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>