drm/i915: Reference objects on the ww lock list
If not referencing object when locking, for more advanced ww transactions,
we need to require all code to be able to unlock their ww locks if we
exit the scope of the object. That leads to complex unlocking rules that
are difficult to follow by programmers and errors the origin of which is
difficult to find, like list corruption. Requiring that we take a reference
when locking will also help pinpoint situations where the lock is taken
when the refcount has reached zero, which is not allowed in TTM since that
prohibits dma_resv individualization and/or early object destruction.

Take a refcount on the object when putting it on a lockint list.
Drawback is an additional atomic operation on lock and unlock.

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>